+++ opentelemetry-exporter-google-cloud/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+- Fix breakages for opentelemetry-python v0.14b0
+  ([#79](https://github.com/GoogleCloudPlatform/opentelemetry-operations-python/pull/79),
+  [#83](https://github.com/GoogleCloudPlatform/opentelemetry-operations-python/pull/83))
+
 ## Version 0.13b0
 
 Released 2020-09-17